Add work in progress XXH3 implementation.
It has no tests yet and the XXH128 hasn't been cleaned up and documented,
so take it for a grain of salt.

Once test values are updated, then the hash can be fully validated.

This follows the version of XXH3 found in the official xxh3 branch, not
dev. This is because there were some major bugs found in xxh3 post release,
and I feel that promoting the work in progress version is better.

Once the XXH3 is properly hardened, this will be merged into master.